230214 김형호

+++ AR41DX/Engine/Include/Resource/Sound/SoundManager.h
@@ -22,8 +22,8 @@ public:
 	bool CreateSoundChannel(const std::string& Name);
 	bool LoadSound(const std::string& GroupName, const std::string& Name,
 		bool Loop, const char* FileName, const std::string& PathName = SOUND_PATH);
-	bool SetVolume(int Volume);
-	bool SetVolume(const std::string& GroupName, int Volume);
+	bool SetVolume(float Volume);
+	bool SetVolume(const std::string& GroupName, float Volume);
 	bool SoundPlay(const std::string& Name);
 	bool SoundStop(const std::string& Name);
 	bool SoundPause(const std::string& Name);
